Add: Refactor and Logger
